Backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
firebase-admin==6.3.0
python-constraint==1.4.0
reportlab==4.0.7
openpyxl==3.1.2
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2